Add libtorus
This is the actual 'library' for the Torus. It isn't a 'true' library
like would be traditionally used in C/++. Instead, since it's a template
data type, it's just a single header file. However, using meson, it
should still be installable to any system that actually wants to use
this datatype. If there exists a single problem that something like this
could solve.