Add unit testing
This adds the framework for unittesting and adding functionality to this
little library. The testing framework is Catch2, which is available on
Ubuntu 22.04 LTS from the repositories. From looking through the
documentation, there appears to be a way for CMAKE to pull Catch2 by
itself and make the header available for testing, but since I have
decided to use Meson instead it does not appear to be an option.

Lastly, this update sets up Meson to start really building and testing
the library. I even added some code to build docs with Doxygen, though
the package is still a few hours work from thorough documentation.