Add reference counting and comparison operation
This adds reference counting to the object to track how many instances
of this Torus exist in the heap. If an operation that overwrites the
curernt object occurs, or if the current object is deleted, it will
decrement the reference counter or delete the Torus based on how many
existing references there are to the current Torus.