Fix pointer issue with references
In my comparisons I was comparing the address of the pointer to an
integer, not actually comparing the value the pointer pointed to to the
integer.

Also, in my incrementing and decrementing of the references, I was
increasing and decreasing the pointers address, not the actual value the
pointer pointed too. This is valuable knowledge to hang on to in the
future.